+++ mama/c_cpp/src/c/log.c
@@ -198,7 +198,7 @@ mamaLog_logLimitReached(void)
 	if(g_logFilePolicy != LOGFILE_UNBOUNDED)
 	{
 		/* Check to see if the log file has reached the maximum size. */
-		if(ftell(gMamaControlledLogFile) > g_maxLogSize)
+		if(ftell(gMamaControlledLogFile) > (long) g_maxLogSize)
 		{
 			/* Upgrade to a writer lock. */
 			MRSW_RESULT lu = MRSWLock_upgrade(g_lock);
@@ -207,7 +207,7 @@ mamaLog_logLimitReached(void)
 				/* Check the file size again under the writer lock in case another
 				 * thread has rolled the file while this one has been waiting.
 				 */
-				if(ftell(gMamaControlledLogFile) > g_maxLogSize)
+				if(ftell(gMamaControlledLogFile) > (long) g_maxLogSize)
 				{
 				printf("Logfile size has been reached.\n");
 				switch (g_logFilePolicy)
